Refactored item package to use BigDecimal for more accurate calculation of prices

+++ src/main/java/items/Items.java
@@ -1,10 +1,12 @@
 package items;
 
+import java.math.BigDecimal;
+
 public interface Items {
 	
 	public Item getNewItem(String itemName);
-	public Double getItemPrice(String itemName);
-	public void addNewItemType(String itemName, Double unitPrice);
-	public void updateItemPrice(String itemName, Double unitPrice);
+	public BigDecimal getItemPrice(String itemName);
+	public void addNewItemType(String itemName, BigDecimal unitPrice);
+	public void updateItemPrice(String itemName, BigDecimal unitPrice);
 
 }
